refactor(web): split usePushNotifications into env + subscription
Previous hook (243 LoC) bundled iOS/standalone detection, permission
state, FCM token CRUD, request-permission flow, foreground onMessage
listener and DB opt-in sync. Split into:

- hooks/push/usePushEnvironment.ts (env + permission state, no firebase)
- hooks/push/usePushSubscription.ts (firebase token work + listeners)

The public hook (`usePushNotifications(userId)`) stays a thin
facade with an unchanged return shape, so call sites remain
untouched. Also types `err: unknown` in the request-permission
catch.
